ManfaaAiAPI/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.12.5
python-dotenv==1.0.1
PyMuPDF==1.26.6
sentence-transformers==5.1.2
scikit-learn==1.7.2
langchain-core==1.2.5
langchain-text-splitters==1.0.0
langchain-chroma==1.0.0
langchain-huggingface==1.1.0
langchain-groq==1.1.1
chromadb==1.3.5
sqlalchemy==2.0.36
pymysql==1.1.2
transformers==4.47.0